main/requirements.txt
opencv-python
numpy
pyserial
requests
keras
tensorflow
Pillow
flask
flask-socketio
eventlet
